Refactor ModuleImporter out of ModuleInstaller
